fix(IndexManager): Refresh before reindex
When performing a migrateLockdown operation there is a copy from the
source index to a temp index, then the source index is recreated and
repopulated from the temp one, but there may not be enough time for the
temp index to finish processing the first _reindex_ operation so the
second _reindex_ operation loses data.

A simple solution is to force a _refresh_ in the temp index before the
copy so that its internal data is up-to-date.