feat(search): Add method for a 'one of' filter
The withOneOfFilter accepts a collection of filters and adds a clause to
search for documents that mataches at least one of the filters in the
group.
This differs from the withGrant option by requiring the client to
provide all the possible criteria at once, so that becomes an isolated
group and won't interfere with other optional clauses.