Bump to version 0.1.0
This version adds a new search option: `withOneOfFilter` filters documents
that match at least one criteria in a group.
This differs from the `withGrant` option by requiring the client to
provide all the possible criteria at once, so that becomes an isolated
group and won't interfere with other optional clauses.